requirements.txt
Pillow==5.2.0
django>=2.0.10
django-ckeditor==5.6.0


django-qr-code==1.0.0
python-dotenv==0.9.1
requests>=2.20.0
urllib3==1.23
